+++ src/fixIcelandicLocale.test.ts
@@ -71,12 +71,8 @@ describe('_PatchedNumberFormat', () => {
     ).toBe('1 t.');
 
     // Certain methods are instance bound and can be called without throwing
-    const { format, formatRange, formatToParts, formatRangeToParts } =
-      new _PatchedNumberFormat('is');
+    const { format } = new _PatchedNumberFormat('is');
     expect(() => format(123)).not.toThrow();
-    expect(() => formatRange(123, 234)).not.toThrow();
-    expect(() => formatToParts(123)).not.toThrow();
-    expect(() => formatRangeToParts(123, 234)).not.toThrow();
   });
 
   test('Has static methods', () => {
@@ -178,12 +174,8 @@ describe('_PatchedDateTimeFormat', () => {
     );
 
     // Certain methods are instance bound and can be called without throwing
-    const { format, formatRange, formatToParts, formatRangeToParts } =
-      new _PatchedDateTimeFormat('is');
+    const { format } = new _PatchedDateTimeFormat('is');
     expect(() => format(d)).not.toThrow();
-    expect(() => formatRange(d, d)).not.toThrow();
-    expect(() => formatToParts(d)).not.toThrow();
-    expect(() => formatRangeToParts(d, d)).not.toThrow();
   });
 
   test('Has static methods', () => {
@@ -410,15 +402,6 @@ describe('_PatchedPluralRules', () => {
     // respects preceding locales ...including "da"
     expect(new _PatchedPluralRules(['en', 'is']).select(21)).toBe('other');
     expect(new _PatchedPluralRules(['da', 'is']).select(21)).toBe('other');
-
-    // Certain methods are instance bound and can be called without throwing
-    const {
-      select,
-      // @ts-expect-error  (TS doesn't know about the .selectRange() method ...yet?)
-      selectRange,
-    } = new _PatchedPluralRules('is');
-    expect(() => select(21)).not.toThrow();
-    expect(() => selectRange(21, 22)).not.toThrow();
   });
 
   test('Has static methods', () => {
@@ -470,11 +453,6 @@ describe('_PatchedListFormat', () => {
     // respects preceding locales ...including "da"
     expect(new _PatchedListFormat(['en', 'is'], opts).format(input)).toBe('a or b');
     expect(new _PatchedListFormat(['da', 'is'], opts).format(input)).toBe('a eller b');
-
-    // Certain methods are instance bound and can be called without throwing
-    const { format, formatToParts } = new _PatchedListFormat('is', opts);
-    expect(() => format(input)).not.toThrow();
-    expect(() => formatToParts(input)).not.toThrow();
   });
 
   test('Has static methods', () => {